Backend/db_helper.py
import mysql.connector
from contextlib import contextmanager
from logging_setup import setup_logger


logger = setup_logger('db_helper')




@contextmanager
def get_db_cursor(commit=False):
    connection = mysql.connector.connect(
        host="localhost",
        user="root",
        password="pgk97@MKVP",
        database="expense_manager"
    )

    cursor = connection.cursor(dictionary=True)
    yield cursor
    if commit:
        connection.commit()
    cursor.close()
    connection.close()

def fetch_expenses_for_date(expense_date):
    logger.info("fetch_expenses_for_date called with {expense_date}")
    with get_db_cursor() as cursor:
        cursor.execute("SELECT * FROM expenses WHERE expense_date = %s", (expense_date,))
        expenses = cursor.fetchall()
    return expenses


def delete_expenses_for_date(expense_date):
    logger.info("fetch_expenses_for_date called with {expense_date}")
    with get_db_cursor(commit=True) as cursor:
        cursor.execute("DELETE FROM expenses WHERE expense_date = %s", (expense_date,))


def insert_expense(expense_date, amount, category, notes):
    logger.info(f"insert_expense called with {expense_date}, amount:{amount}, category:{category}, notes:{notes}")
    with get_db_cursor(commit=True) as cursor:
        cursor.execute(
            "INSERT INTO expenses (expense_date, amount, category, notes) VALUES (%s, %s, %s, %s)",
            (expense_date, amount, category, notes)
        )

def fetch_expense_summary(start_date, end_date):  # Added colon
    logger.info(f"fetch_expense_summary called with {start_date} end: {end_date}")
    with get_db_cursor() as cursor:  # Added parentheses ()
        cursor.execute(
            '''SELECT category, SUM(amount) as total
               FROM expenses WHERE expense_date
               BETWEEN %s AND %s
               GROUP BY category''',  # Added comma, removed semicolon
            (start_date, end_date)
        )
        data = cursor.fetchall()
    return data


if __name__ == "__main__":

    expenses = fetch_expenses_for_date("2024-09-30")
    print(expenses)
    insert_expense("2024-08-25", 40, "Food", "Eat tasty samossa chat")
    #delete_expenses_for_date("2024-08-25")
    summary = fetch_expense_summary("2024-08-01", "2024-08-05")  # Fixed function name
    for record in summary:
        print(record)