requirements.txt
streamlit==1.53.1
pandas==2.3.3
requests==2.32.5
fastapi==0.128.0
uvicorn==0.40.0
mysql-connector-python==9.5.0
pytest== 9.0.2
pydantic==2.12.5